refactor(guildDomain): split GuildDomainMenu.ts (801L) into Shared/Views/Menu
- GuildDomainShared.ts (186L): types, constants, navigation/upgrade helpers, getBuildingSummary.
- GuildDomainViews.ts (348L): all buildXxxContainer + addFoodSections.
- GuildDomainMenu.ts (266L): handlers (upgrade/foodBuy/xpBuy), collectors, register, public API.

Also fix latent bug: FOOD_KEYS now matches PetConstants.PET_FOOD_BY_ID order
(common, herbivorous, carnivorous, ultimate) so data.food[foodKey] returns the right stock.
Same fix in pantry container i18n params.
