+++ Discord/src/commands/player/report/guildDomain/GuildDomainViews.ts
@@ -1,0 +1,348 @@
+import {
+	ButtonBuilder, ButtonStyle,
+	ContainerBuilder,
+	SectionBuilder, SeparatorBuilder, SeparatorSpacingSize,
+	TextDisplayBuilder
+} from "discord.js";
+import i18n from "../../../../translations/i18n";
+import { CrowniclesIcons } from "../../../../../../Lib/src/CrowniclesIcons";
+import { ReportCityMenuIds } from "../ReportCityMenuConstants";
+import { Language } from "../../../../../../Lib/src/Language";
+import {
+	GuildBuilding, GuildDomainConstants
+} from "../../../../../../Lib/src/constants/GuildDomainConstants";
+import { PetConstants } from "../../../../../../Lib/src/constants/PetConstants";
+import {
+	addDomainNavigation, addStatusMessage, addUpgradeSection,
+	BUILDING_ICONS, FOOD_KEYS, getBuildingSummary,
+	GuildDomainData, GuildDomainMenuContext
+} from "./GuildDomainShared";
+
+export function buildMainDomainContainer(ctx: GuildDomainMenuContext, statusMessage?: string): ContainerBuilder {
+	const {
+		data, lng, pseudo
+	} = ctx;
+	const container = new ContainerBuilder();
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			`### ${i18n.t("commands:report.city.guildDomain.menuTitle", {
+				lng, pseudo, guildName: data.guildName
+			})}`
+		)
+	);
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			i18n.t("commands:report.city.guildDomain.menuIntro", {
+				lng, guildName: data.guildName
+			})
+		)
+	);
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			i18n.t("commands:report.city.guildDomain.guildLevelAndTreasury", {
+				lng,
+				guildLevel: data.guildLevel,
+				treasury: data.treasury,
+				playerMoney: data.playerMoney
+			})
+		)
+	);
+
+	container.addSeparatorComponents(new SeparatorBuilder().setSpacing(SeparatorSpacingSize.Small));
+
+	for (const building of Object.values(GuildBuilding)) {
+		const currentLevel = data[`${building}Level` as keyof typeof data] as number;
+		const maxLevel = GuildDomainConstants.BUILDINGS[building].maxLevel;
+		const buildingName = i18n.t(`commands:report.city.guildDomain.buildings.${building}`, { lng });
+		const description = getBuildingSummary(building, currentLevel, data, lng);
+		const buildingIcon = BUILDING_ICONS[building];
+
+		container.addSectionComponents(
+			new SectionBuilder()
+				.addTextDisplayComponents(
+					new TextDisplayBuilder().setContent(
+						`${buildingIcon} **${buildingName}** — ${i18n.t("commands:report.city.guildDomain.levelDisplay", {
+							lng, level: currentLevel, maxLevel
+						})}\n${description}`
+					)
+				)
+				.setButtonAccessory(
+					new ButtonBuilder()
+						.setCustomId(`${ReportCityMenuIds.GUILD_DOMAIN_ENTER_PREFIX}${building}`)
+						.setLabel(i18n.t(`commands:report.city.guildDomain.enterBuilding.${building}`, { lng }))
+						.setStyle(ButtonStyle.Primary)
+				)
+		);
+	}
+
+	addStatusMessage(container, statusMessage);
+
+	addDomainNavigation(
+		container,
+		ctx,
+		i18n.t("commands:report.city.guildDomain.leaveDomain", { lng }),
+		ReportCityMenuIds.BACK_TO_CITY
+	);
+
+	return container;
+}
+
+function addFoodSections(container: ContainerBuilder, data: GuildDomainData, lng: Language): void {
+	const foodTypes = PetConstants.PET_FOOD_BY_ID;
+
+	for (let i = 0; i < foodTypes.length; i++) {
+		const foodType = foodTypes[i];
+		const foodKey = FOOD_KEYS[i];
+		const currentStock = data.food[foodKey];
+		const cap = data.foodCaps[i];
+		const remainingSlots = cap - currentStock;
+		const price = GuildDomainConstants.SHOP_PRICES.FOOD[i];
+		const maxAffordable = Math.floor(data.playerMoney / price);
+		const maxBuyable = Math.min(remainingSlots, maxAffordable);
+		const foodName = i18n.t(`models:foods.${foodType}`, {
+			lng, count: 1
+		});
+		const foodEmoji = CrowniclesIcons.foods[foodType] ?? "";
+
+		container.addSectionComponents(
+			new SectionBuilder()
+				.addTextDisplayComponents(
+					new TextDisplayBuilder().setContent(
+						`${foodEmoji} **${foodName}** — ${currentStock}/${cap}\n*${price} ${CrowniclesIcons.unitValues.money} l'unité*`
+					)
+				)
+				.setButtonAccessory(
+					new ButtonBuilder()
+						.setCustomId(`${ReportCityMenuIds.GUILD_DOMAIN_SHOP_FOOD_PREFIX}${foodType}_${Math.max(maxBuyable, 1)}`)
+						.setLabel(i18n.t("commands:report.city.guildDomain.subMenus.shop.buyFoodButton", {
+							lng,
+							amount: Math.max(maxBuyable, 1),
+							food: foodName,
+							cost: price * Math.max(maxBuyable, 1)
+						}))
+						.setStyle(ButtonStyle.Primary)
+						.setDisabled(maxBuyable <= 0)
+				)
+		);
+	}
+}
+
+export function buildShopContainer(ctx: GuildDomainMenuContext, statusMessage?: string): ContainerBuilder {
+	const {
+		data, lng
+	} = ctx;
+	const container = new ContainerBuilder();
+	const shopLevel = data.shopLevel;
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			`### ${i18n.t("commands:report.city.guildDomain.subMenus.shop.title", { lng })}`
+		)
+	);
+
+	if (shopLevel === 0) {
+		container.addTextDisplayComponents(
+			new TextDisplayBuilder().setContent(
+				i18n.t("commands:report.city.guildDomain.subMenus.shop.notBuilt", { lng })
+			)
+		);
+
+		container.addSeparatorComponents(new SeparatorBuilder().setSpacing(SeparatorSpacingSize.Small));
+		addUpgradeSection(container, GuildBuilding.SHOP, ctx);
+		addStatusMessage(container, statusMessage);
+		addDomainNavigation(container, ctx, i18n.t("commands:report.city.guildDomain.backToDomain", { lng }), ReportCityMenuIds.GUILD_DOMAIN_BACK);
+		return container;
+	}
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			i18n.t("commands:report.city.guildDomain.subMenus.shop.description", {
+				lng, playerMoney: data.playerMoney
+			})
+		)
+	);
+
+	container.addSeparatorComponents(new SeparatorBuilder().setSpacing(SeparatorSpacingSize.Small));
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			i18n.t("commands:report.city.guildDomain.subMenus.shop.buyFoodLabel", { lng })
+		)
+	);
+
+	addFoodSections(container, data, lng);
+
+	container.addSeparatorComponents(new SeparatorBuilder().setSpacing(SeparatorSpacingSize.Small));
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			i18n.t("commands:report.city.guildDomain.subMenus.shop.buyXpLabel", { lng })
+		)
+	);
+
+	container.addSectionComponents(
+		new SectionBuilder()
+			.addTextDisplayComponents(
+				new TextDisplayBuilder().setContent(
+					`${CrowniclesIcons.shopItems.smallGuildXp} **${i18n.t("commands:report.city.guildDomain.subMenus.shop.buyXpSmall", {
+						lng, cost: GuildDomainConstants.SHOP_PRICES.SMALL_XP
+					})}**\n${i18n.t("commands:report.city.guildDomain.subMenus.shop.buyXpSmallDescription", { lng })}`
+				)
+			)
+			.setButtonAccessory(
+				new ButtonBuilder()
+					.setCustomId(`${ReportCityMenuIds.GUILD_DOMAIN_SHOP_XP_PREFIX}small`)
+					.setLabel(i18n.t("commands:report.city.guildDomain.subMenus.shop.buyXpSmall", {
+						lng, cost: GuildDomainConstants.SHOP_PRICES.SMALL_XP
+					}))
+					.setStyle(ButtonStyle.Success)
+					.setDisabled(data.playerMoney < GuildDomainConstants.SHOP_PRICES.SMALL_XP)
+			)
+	);
+
+	container.addSectionComponents(
+		new SectionBuilder()
+			.addTextDisplayComponents(
+				new TextDisplayBuilder().setContent(
+					`${CrowniclesIcons.shopItems.bigGuildXp} **${i18n.t("commands:report.city.guildDomain.subMenus.shop.buyXpBig", {
+						lng, cost: GuildDomainConstants.SHOP_PRICES.BIG_XP
+					})}**\n${i18n.t("commands:report.city.guildDomain.subMenus.shop.buyXpBigDescription", { lng })}`
+				)
+			)
+			.setButtonAccessory(
+				new ButtonBuilder()
+					.setCustomId(`${ReportCityMenuIds.GUILD_DOMAIN_SHOP_XP_PREFIX}big`)
+					.setLabel(i18n.t("commands:report.city.guildDomain.subMenus.shop.buyXpBig", {
+						lng, cost: GuildDomainConstants.SHOP_PRICES.BIG_XP
+					}))
+					.setStyle(ButtonStyle.Success)
+					.setDisabled(data.playerMoney < GuildDomainConstants.SHOP_PRICES.BIG_XP)
+			)
+	);
+
+	container.addSeparatorComponents(new SeparatorBuilder().setSpacing(SeparatorSpacingSize.Small));
+	addUpgradeSection(container, GuildBuilding.SHOP, ctx);
+	addStatusMessage(container, statusMessage);
+	addDomainNavigation(container, ctx, i18n.t("commands:report.city.guildDomain.backToDomain", { lng }), ReportCityMenuIds.GUILD_DOMAIN_BACK);
+
+	return container;
+}
+
+export function buildShelterContainer(ctx: GuildDomainMenuContext, statusMessage?: string): ContainerBuilder {
+	const {
+		data, lng
+	} = ctx;
+	const container = new ContainerBuilder();
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			`### ${i18n.t("commands:report.city.guildDomain.subMenus.shelter.title", { lng })}`
+		)
+	);
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			i18n.t("commands:report.city.guildDomain.subMenus.shelter.description", {
+				lng,
+				slots: GuildDomainConstants.getShelterSlots(data.shelterLevel)
+			})
+		)
+	);
+
+	container.addSeparatorComponents(new SeparatorBuilder().setSpacing(SeparatorSpacingSize.Small));
+	addUpgradeSection(container, GuildBuilding.SHELTER, ctx);
+	addStatusMessage(container, statusMessage);
+	addDomainNavigation(container, ctx, i18n.t("commands:report.city.guildDomain.backToDomain", { lng }), ReportCityMenuIds.GUILD_DOMAIN_BACK);
+
+	return container;
+}
+
+export function buildPantryContainer(ctx: GuildDomainMenuContext, statusMessage?: string): ContainerBuilder {
+	const {
+		data, lng
+	} = ctx;
+	const container = new ContainerBuilder();
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			`### ${i18n.t("commands:report.city.guildDomain.subMenus.pantry.title", { lng })}`
+		)
+	);
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			i18n.t("commands:report.city.guildDomain.subMenus.pantry.description", { lng })
+		)
+	);
+
+	container.addSeparatorComponents(new SeparatorBuilder().setSpacing(SeparatorSpacingSize.Small));
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			i18n.t("commands:report.city.guildDomain.foodInfo", {
+				lng,
+				common: data.food.common,
+				commonCap: data.foodCaps[0],
+				herbivorous: data.food.herbivorous,
+				herbivorousCap: data.foodCaps[1],
+				carnivorous: data.food.carnivorous,
+				carnivorousCap: data.foodCaps[2],
+				ultimate: data.food.ultimate,
+				ultimateCap: data.foodCaps[3]
+			})
+		)
+	);
+
+	container.addSeparatorComponents(new SeparatorBuilder().setSpacing(SeparatorSpacingSize.Small));
+	addUpgradeSection(container, GuildBuilding.PANTRY, ctx);
+	addStatusMessage(container, statusMessage);
+	addDomainNavigation(container, ctx, i18n.t("commands:report.city.guildDomain.backToDomain", { lng }), ReportCityMenuIds.GUILD_DOMAIN_BACK);
+
+	return container;
+}
+
+export function buildTrainingGroundContainer(ctx: GuildDomainMenuContext, statusMessage?: string): ContainerBuilder {
+	const {
+		data, lng
+	} = ctx;
+	const container = new ContainerBuilder();
+	const love = GuildDomainConstants.getTrainingLovePerDay(data.trainingGroundLevel);
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			`### ${i18n.t("commands:report.city.guildDomain.subMenus.training.title", { lng })}`
+		)
+	);
+
+	container.addTextDisplayComponents(
+		new TextDisplayBuilder().setContent(
+			love === 0
+				? i18n.t("commands:report.city.guildDomain.subMenus.training.descriptionInactive", { lng })
+				: i18n.t("commands:report.city.guildDomain.subMenus.training.descriptionActive", {
+					lng, love
+				})
+		)
+	);
+
+	container.addSeparatorComponents(new SeparatorBuilder().setSpacing(SeparatorSpacingSize.Small));
+	addUpgradeSection(container, GuildBuilding.TRAINING_GROUND, ctx);
+	addStatusMessage(container, statusMessage);
+	addDomainNavigation(container, ctx, i18n.t("commands:report.city.guildDomain.backToDomain", { lng }), ReportCityMenuIds.GUILD_DOMAIN_BACK);
+
+	return container;
+}
+
+export function buildBuildingContainer(building: GuildBuilding, ctx: GuildDomainMenuContext, statusMessage?: string): ContainerBuilder {
+	switch (building) {
+		case GuildBuilding.SHOP:
+			return buildShopContainer(ctx, statusMessage);
+		case GuildBuilding.SHELTER:
+			return buildShelterContainer(ctx, statusMessage);
+		case GuildBuilding.PANTRY:
+			return buildPantryContainer(ctx, statusMessage);
+		case GuildBuilding.TRAINING_GROUND:
+			return buildTrainingGroundContainer(ctx, statusMessage);
+		default:
+			return new ContainerBuilder();
+	}
+}
